+++ src/main/java/nextstep/favorite/ui/FavoriteController.java
@@ -5,7 +5,7 @@ import nextstep.favorite.application.DeleteFavoriteService;
 import nextstep.favorite.application.GetFavoriteService;
 import nextstep.favorite.application.dto.FavoriteRequest;
 import nextstep.favorite.application.dto.FavoriteResponse;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.domain.LoginMember;
 import nextstep.auth.ui.AuthenticationPrincipal;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
