+++ src/test/java/nextstep/auth/acceptance/AuthAcceptanceTest.java
@@ -3,7 +3,7 @@ package nextstep.auth.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.member.application.dto.GitHubLoginRequest;
+import nextstep.auth.application.dto.GitHubLoginRequest;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import nextstep.utils.AcceptanceTest;
