+++ src/test/java/nextstep/auth/ui/GitHubAccessTokenTestController.java
@@ -1,8 +1,8 @@
 package nextstep.auth.ui;
 
-import nextstep.member.application.dto.GitHubAccessTokenRequest;
-import nextstep.member.application.dto.GitHubAccessTokenResponse;
-import nextstep.member.application.dto.GitHubProfileResponse;
+import nextstep.auth.application.dto.GitHubAccessTokenRequest;
+import nextstep.auth.application.dto.GitHubAccessTokenResponse;
+import nextstep.auth.application.dto.GitHubProfileResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
