+++ src/test/java/nextstep/favorite/fixture/MemberFixture.java
@@ -4,7 +4,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.member.application.dto.MemberRequest;
-import nextstep.member.application.dto.TokenRequest;
+import nextstep.auth.application.dto.TokenRequest;
 import org.springframework.http.MediaType;
 
 import static org.assertj.core.api.Assertions.assertThat;
