+++ src/test/java/nextstep/favorite/unit/GetFavoriteTest.java
@@ -4,7 +4,7 @@ import nextstep.favorite.application.GetFavoriteService;
 import nextstep.favorite.application.dto.FavoriteResponse;
 import nextstep.favorite.domain.Favorite;
 import nextstep.favorite.domain.FavoriteRepository;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import nextstep.subway.line.Line;
